Use React state to update plots
